+++ packages/framework/src/Model/Product/Search/Export/ProductSearchExportCronModule.php
@@ -3,69 +3,25 @@
 namespace Shopsys\FrameworkBundle\Model\Product\Search\Export;
 
 use Shopsys\FrameworkBundle\Component\Domain\Domain;
+use Shopsys\FrameworkBundle\Component\Elasticsearch\AbstractExportCronModule;
 use Shopsys\FrameworkBundle\Component\Elasticsearch\IndexDefinitionLoader;
 use Shopsys\FrameworkBundle\Component\Elasticsearch\IndexFacade;
-use Shopsys\FrameworkBundle\Component\Elasticsearch\IndexRegistry;
 use Shopsys\FrameworkBundle\Model\Product\Elasticsearch\ProductIndex;
-use Shopsys\Plugin\Cron\SimpleCronModuleInterface;
-use Symfony\Bridge\Monolog\Logger;
-use Symfony\Component\Console\Output\NullOutput;
 
-class ProductSearchExportCronModule implements SimpleCronModuleInterface
+class ProductSearchExportCronModule extends AbstractExportCronModule
 {
     /**
-     * @var \Shopsys\FrameworkBundle\Component\Elasticsearch\IndexFacade
-     */
-    protected $indexFacade;
-
-    /**
-     * @var \Shopsys\FrameworkBundle\Component\Elasticsearch\IndexRegistry
-     */
-    protected $indexRegistry;
-
-    /**
-     * @var \Shopsys\FrameworkBundle\Component\Elasticsearch\IndexDefinitionLoader
-     */
-    protected $indexDefinitionLoader;
-
-    /**
-     * @var \Shopsys\FrameworkBundle\Component\Domain\Domain
-     */
-    protected $domain;
-
-    /**
-     * ProductSearchExportCronModule constructor.
-     *
+     * @param \Shopsys\FrameworkBundle\Model\Product\Elasticsearch\ProductIndex $index
      * @param \Shopsys\FrameworkBundle\Component\Elasticsearch\IndexFacade $indexFacade
-     * @param \Shopsys\FrameworkBundle\Component\Elasticsearch\IndexRegistry $indexRegistry
      * @param \Shopsys\FrameworkBundle\Component\Elasticsearch\IndexDefinitionLoader $indexDefinitionLoader
      * @param \Shopsys\FrameworkBundle\Component\Domain\Domain $domain
      */
     public function __construct(
+        ProductIndex $index,
         IndexFacade $indexFacade,
-        IndexRegistry $indexRegistry,
         IndexDefinitionLoader $indexDefinitionLoader,
         Domain $domain
     ) {
-        $this->indexFacade = $indexFacade;
-        $this->indexRegistry = $indexRegistry;
-        $this->indexDefinitionLoader = $indexDefinitionLoader;
-        $this->domain = $domain;
-    }
-
-    /**
-     * @param \Symfony\Bridge\Monolog\Logger $logger
-     */
-    public function setLogger(Logger $logger)
-    {
-    }
-
-    public function run()
-    {
-        $index = $this->indexRegistry->getIndexByIndexName(ProductIndex::INDEX_NAME);
-        foreach ($this->domain->getAllIds() as $domainId) {
-            $indexDefinition = $this->indexDefinitionLoader->getIndexDefinition($index, $domainId);
-            $this->indexFacade->exportByIndexDefinition($indexDefinition, new NullOutput());
-        }
+        parent::__construct($index, $indexFacade, $indexDefinitionLoader, $domain);
     }
 }
